+++ SDK Submissions/ChimoneyDotNet/ChimoneyDotNet/Models/Account/TransactionDetail.cs
@@ -9,7 +9,7 @@ public class TransactionDetail : Transaction
     public string Email { get; set; }
     public decimal Chimoney { get; set; }
     public decimal ValueInUSD { get; set; }
-    public string[][] EnabledToRedeem { get; set; }
+    public string[] EnabledToRedeem { get; set; }
     //HACK another way to do this?
     public string? Error { get; set; }
 }
